server/requirements.txt
Flask
flask-cors
python-dotenv
# openai
# anthropic
google-generativeai
